requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
PyJWT==2.8.0
SQLAlchemy
psycopg2